finish "blend": can draw opaque and transparent objects in the same time can set blendFunc,blendEquation Material,QuadCommand add blend,blendSrc,blendDst,blendEquation attribute Stage->_camera change to public attribute "camera" Color add alpha attribute Geometry->colors get alpha from material->color
WebGLRenderer:
When drawing a scene with non-transparent and transparent objects the general outline is usually as follows:

Draw all opaque objects first.
Sort all the transparent objects.
Draw all the transparent objects in sorted order.

the way sorting the transparent objects is to retrieve the distance of an object from the viewer's perspective(camera->z - object->z)