refactor: texture asset save the default value, then copy it to texture (not pass cubemap unit test)
this will cause duplicate definition of the attri(like source,width), it can be solved by retain "asset" reference to the TextureAsset in Texture, and Texture and user use "texture.asset.xxx" to set the attri.
but there are two problem:
1.user shouldn't know texture's asset
2.Texture may need overwrite the attri(like CompressedTexture need overwrite sourceRegionMethod getter), but in this way, it can't directly overwrite the attri's getter, instead Texture may add "setXXX" method(like setSourceRegionMethod) and user should invoke "texture.setXXX" to set it.

this will not be consistence that user should use "texture.asset.xxx" or "texture.setXXX" to set texture's attribute.

so here texture asset and texture all define the attri.