fix "if set gameObject's transform before building its bounding region, it will collision wrong" bug: description use bellow unit test code: 	    //box1, box2 is at zero point, and its width,height,depth are all 5
            box1.transform.translate(2, 0, 0);

            director._init();

            box1.transform.translate(6, 0, 0);
            box2.transform.translate(-2, 0, 0);

            director._run(1);

            judgeCollide();	//fail!not collide

here when invoke "_run", the distance between box1 and box2 is 10, so it should collide, but the actual is not!

reason
because the aabb is build from gameObject in "_init" that it's already transformed by "box1.transform.translate(2, 0, 0);".
here box1->aabb transform "translate(2, 0, 0)" twice!
so the actual distance between box1 and box2 is 12 > 10, so not collide

solution
build aabb from geometry's vertices that are not transformed