fix instance_octree_shadow_layer sample->"if not show sphere2(whose shadow layer is diferent from the instance), it's dark" bug: reason if no objects render in one shadow map(not build shadow map), the shadow map will still be drawed and the visibility from the shadow map is all 0.0.
solution
first, i try to not bind and send the shadow map which is not build.
but it fails because of the glsl limit(can't change the shadow map count in "for")

so i modify totalShadowMap_fragment->unpackDepth function, add:
    if(rgbaDepth == vec4(0.0)){
        return 100000.0;
    }

so it can make sure that the visibility from the shadow map is all 1.0.