finish component->Camera add "clone" method
add "requireGetterAndSetter", "ensureGetterAndSetter" decorator

modify FlyCameraController: now support modify "moveSpeed","rotateSpeed", "zoomSpeed"(only when the camera is PerspectiveCamera)

fix "clone decorator" bug:
description
the attr of sibling classes will affect each other!

reason
the "cloneAttributeMembers" is existed in the parent of the class and its children classes share the "cloneAttributeMembers"!

(the "cloneAttributeMembers" isn't existed in the instance because it is add to the data at build time.)

solution
judge to exclude the member of sibling classes
(but it still may be wrong in the case: the sibling classes has the same attri that need be cloned.
in this case, it should mannually define the "clone" method)